test: ServiceUseCase와 Service 테스트를 분리 (#23)
